test: move locale test in separate file
If in the same file as the other, it fails because d2 is a singleton.
